+++ packages/brick_offline_first_with_supabase/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 1.1.1
+
 - Allow a generic type argument for `OfflineFirstWithSupabaseRepository`
 
 ## 1.1.0