+++ packages/brick_supabase_generators/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## 1.0.2
+
+- Interpret `@Supabase(query:)` in `SupabaseFields`
+
 ## 1.0.1
 
 - Apply `@Supabase(foreignKey:)` to `RuntimeSupabaseColumnDefinition#foreignKey`